Create `display_per_host_statistics` in separate file
